backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
cffi==1.17.1
clerk-backend-api==3.0.3
click==8.2.1
cryptography==44.0.3
distro==1.9.0
fastapi==0.115.14
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.93.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pyjwt==2.10.1
python-dotenv==1.1.1
ruff==0.12.1
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.46.2
tqdm==4.67.1
typing-extensions==4.14.0
typing-inspection==0.4.1
uvicorn==0.35.0
